Split file `Clickable`
