+++ src/components/ProblemAdministration/ProblemAdministration.tsx
@@ -9,7 +9,8 @@ import {DropzoneOptions, useDropzone} from 'react-dropzone'
 import {ProblemWithSolutions, SolutionAdministration} from '@/types/api/competition'
 import {useHasPermissions} from '@/utils/useHasPermissions'
 
-import {Button, Link} from '../Clickable/Clickable'
+import {Button} from '../Clickable/Button'
+import {Link} from '../Clickable/Link'
 import {FileDropZone} from '../FileDropZone/FileDropZone'
 import {FileUploader} from '../FileUploader/FileUploader'
 import {Latex} from '../Latex/Latex'
