+++ src/components/Problems/Problems.tsx
@@ -4,7 +4,8 @@ import {useRouter} from 'next/router'
 import {FC, useEffect, useState} from 'react'
 import {useInterval} from 'usehooks-ts'
 
-import {Button, Link} from '@/components/Clickable/Clickable'
+import {Button} from '@/components/Clickable/Button'
+import {Link} from '@/components/Clickable/Link'
 import {SeriesWithProblems} from '@/types/api/competition'
 import {Profile} from '@/types/api/personal'
 import {AuthContainer} from '@/utils/AuthContainer'
