+++ src/pages/test/index.tsx
@@ -1,6 +1,6 @@
 import {NextPage} from 'next'
 
-import {Link} from '@/components/Clickable/Clickable'
+import {Link} from '@/components/Clickable/Link'
 import {PageLayout} from '@/components/PageLayout/PageLayout'
 import {PagePlaceholder} from '@/components/PagePlaceholder/PagePlaceholder'
